Fix comparisons in DittoTrace
Make sure that both variables of a comparison are of the same type.

Fix: 06b52cf06a63 ("Implement DittoTrace")
Bug: 200543075
Test: mm

+++ dittotrace.cpp
@@ -56,8 +56,8 @@ Syscall ProcessLine(const std::string& line) {
   syscall.return_value = line.substr(line.find(")"));
   syscall.return_value = syscall.return_value.substr(syscall.return_value.find("= ") + 2);
 
-  int next = 0;
-  int last = 0;
+  size_t next = 0;
+  size_t last = 0;
   while ((next = raw_arguments.find(", ", last)) != std::string::npos) {
     std::string part = raw_arguments.substr(last, next - last);
     last = next + 2;
@@ -89,7 +89,7 @@ std::map<int, std::vector<Syscall>> ProcessLines(
   std::map<int, std::vector<Syscall>> processed_syscalls_by_pid;
 
   for (const auto& [pid, lines] : lines_by_pid) {
-    for (int i = 0; i < lines.size(); ++i) {
+    for (unsigned int i = 0; i < lines.size(); ++i) {
       auto line = lines[i];
 
       // If only the resumed part of the syscall was found, ignore it
